requirements.txt
aniso8601==3.0.2
atomicwrites==1.3.0
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
coverage==4.5.2
Django==2.1.7
django-cors-headers==2.4.0
django-filter==2.1.0
django-redis==4.10.0
djangorestframework==3.9.2
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
gunicorn==19.9.0
idna==2.8
more-itertools==5.0.0
mysqlclient==1.3.14
pluggy==0.8.1
pprintpp==0.4.0
promise==2.2.1
py==1.7.0
pycountry==18.12.8
pycountry-convert==0.7.2
pytest==4.2.0
pytest-cov==2.6.1
pytest-mock==1.10.1
python-dotenv==0.10.1
pytz==2018.9
redis==3.2.0
repoze.lru==0.7
requests==2.21.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
style==1.1.0
update==0.0.1
urllib3==1.24.1
